Fixed add liquidity approvals
